indexeddb: Move the state store migrations to a separate file
